Modificati i file dei test e il file della lsa
