+++ src/content_analyzer/embedding_learner/embedding_learner.py
@@ -2,7 +2,6 @@ import time
 from abc import ABC, abstractmethod
 from typing import List
 
-
 from src.content_analyzer.information_processor.information_processor import TextProcessor
 from src.content_analyzer.information_processor.nlp import NLTK
 from src.content_analyzer.raw_information_source import RawInformationSource
