+++ src/content_analyzer/embedding_learner/latent_semantic_analysis.py
@@ -24,27 +24,24 @@ class GensimLatentSemanticAnalysis(EmbeddingLearner):
                 "preprocessor = " + str(self.__preprocessor) + " >"
 
     @staticmethod
-    def create_dictionary(docs) -> Dictionary:
+    def __create_dictionary(docs) -> Dictionary:
         return Dictionary(docs)
 
     @staticmethod
-    def create_word_docs_matrix(docs, dictionary) -> List[str]:
+    def __create_word_docs_matrix(docs, dictionary) -> List[str]:
         return [dictionary.doc2bow(doc) for doc in docs]
 
     def fit(self):
         """
-        Creates the model for the embedding
+        Creates the model for the embedding and sets it as the value of the 'model' attribute
 
         Args:
             dictionary: Can be either a stream of document vectors or a sparse matrix of shape
                 (number_of_terms, number_of_documents)
             word_docs_matrix: Mapping between the IDs and the words
 
-        Returns:
-            LsiModel: The LSI (Latent Semantic Index) built.
         """
         docs = self.extract_corpus()
-        dictionary = GensimLatentSemanticAnalysis.create_dictionary(docs)
-        word_docs_matrix = GensimLatentSemanticAnalysis.create_word_docs_matrix(docs, dictionary)
-        return LsiModel(word_docs_matrix, id2word=dictionary)
-
+        dictionary = GensimLatentSemanticAnalysis.__create_dictionary(docs)
+        word_docs_matrix = GensimLatentSemanticAnalysis.__create_word_docs_matrix(docs, dictionary)
+        self.set_model(LsiModel(word_docs_matrix, id2word=dictionary))
